a_useless_button: added example.
