+++ examples/a_useless_button/Makefile
@@ -1,7 +1,7 @@
 all : flash
 
-TARGET:=a_useless_code
-ADDITIONAL_C_FILES+=fun_timer.h fun_button.h
+TARGET:=a_useless_button
+ADDITIONAL_C_FILES+=fun_timer.h fun_button.h fun_gpio.h
 
 include ../../ch32v003fun/ch32v003fun.mk
