+++ examples/a_useless_button/a_useless_button.c
@@ -1,6 +1,9 @@
 /*
  * Example of a useless code. It doesn't do anything good except being great at what it does.
  * It demonstrates the power of repetition: if you repeat the samething many times over, it becomes you.
+ * 
+ * fun_gpio.h is based on ch32v_hal.h by Larry Bank
+ * fun_timer.h is based on Example SysTick with IRQs work of E. Brombaugh and C. Lohr
  * by unicab369
  */
 
@@ -8,22 +11,22 @@
 #include <stdio.h>
 #include "fun_button.h"
 
+#define BUTTON_PIN 0xC0
+#define LED_PIN 0xD0
+
 void button_onSingleClick() {
    printf("\nI'M USELESS.");
-	digitalWrite(0xC0, 0);
+	digitalWrite(LED_PIN, !digitalRead(LED_PIN));
 }
 
 void button_onDoubleClick() {
    printf("\nI'M USELESS TWICE.");
-	digitalWrite(0xC0, !digitalRead(0xC0));
 }
 
 int main()
 {  
-	pinMode(0xC0, OUTPUT);
-	digitalWrite(0xC0, 1);
-
-	button_setup(0xC3);
+	pinMode(LED_PIN, OUTPUT);
+	button_setup(BUTTON_PIN);
 
 	printf("looping...\n\r");
 	while(1)
