+++ examples/a_useless_button/fun_timer.h
@@ -1,0 +1,74 @@
+/*
+ * Copied and Adapt from Example SysTick with IRQs work of E. Brombaugh and C. Lohr
+ */
+
+#include "ch32v003fun.h"
+#include <stdio.h>
+
+/* some bit definitions for systick regs */
+#define SYSTICK_SR_CNTIF (1<<0)
+#define SYSTICK_CTLR_STE (1<<0)
+#define SYSTICK_CTLR_STIE (1<<1)
+#define SYSTICK_CTLR_STCLK (1<<2)
+#define SYSTICK_CTLR_STRE (1<<3)
+#define SYSTICK_CTLR_SWIE (1<<31)
+
+volatile uint32_t systick_cnt;
+
+/*
+ * Start up the SysTick IRQ
+ */
+void systick_init(void)
+{
+	/* disable default SysTick behavior */
+	SysTick->CTLR = 0;
+	
+	/* enable the SysTick IRQ */
+	NVIC_EnableIRQ(SysTicK_IRQn);
+	
+	/* Set the tick interval to 1ms for normal op */
+	SysTick->CMP = (FUNCONF_SYSTEM_CORE_CLOCK/1000)-1;
+	
+	/* Start at zero */
+	SysTick->CNT = 0;
+	systick_cnt = 0;
+	
+	/* Enable SysTick counter, IRQ, HCLK/1 */
+	SysTick->CTLR = SYSTICK_CTLR_STE | SYSTICK_CTLR_STIE |
+					SYSTICK_CTLR_STCLK;
+}
+
+/*
+ * SysTick ISR just counts ticks
+ * note - the __attribute__((interrupt)) syntax is crucial!
+ */
+void SysTick_Handler(void) __attribute__((interrupt));
+void SysTick_Handler(void)
+{
+	// move the compare further ahead in time.
+	// as a warning, if more than this length of time
+	// passes before triggering, you may miss your
+	// interrupt.
+	SysTick->CMP += (FUNCONF_SYSTEM_CORE_CLOCK/1000);
+
+	/* clear IRQ */
+	SysTick->SR = 0;
+
+	/* update counter */
+	systick_cnt++;
+}
+
+void timer_setup() {
+	// init systick @ 1ms rate
+	printf("initializing systick...");
+	systick_init();
+	printf("done.\n\r");
+}
+
+uint32_t timer_getTick() {
+	return systick_cnt;
+}
+
+uint32_t timer_getCount() {
+	return SysTick->CNT;
+}
