+++ examples/a_useless_button/funconfig.h
