--- examples/a_useless_code/README.md
@@ -1,3 +1,0 @@
-# A Useless Code
-Example of a useless code. It doesn't do anything good except being great at what it does.
-It demonstrates the power of repetition: if you repeat the samething many times over, it becomes you.
