+++ platformio.ini
@@ -27,8 +27,8 @@ extra_libs_srcs = +<extralibs>
 ; 3. Add additional build flags as needed (see uartdemo)
 ; 4. Switch to new environment in VSCode bottom taskbar (https://docs.platformio.org/en/latest/integration/ide/vscode.html#project-tasks)
 
-[env:a_useless_code]
-build_src_filter = ${fun_base.build_src_filter} +<examples/a_useless_code>
+[env:a_useless_button]
+build_src_filter = ${fun_base.build_src_filter} +<examples/a_useless_button>
 
 [env:adc_dma_opamp]
 build_src_filter = ${fun_base.build_src_filter} +<examples/adc_dma_opamp>
